Add new plugin for Ganglia
now, we add a plugin called ganglia, it can send collectd's metrics
to ganglia(accurate, is gmond), and then we can see the metrics in
ganglia system. In order to use the plugin, we need to add some lines
in collectd.conf, like:
LoadPlugin ganglia
<Plugin "ganglia">
  GmondConfPath "/etc/ganglia/gmond.conf"
</Plugin>
So, we need to load the plugin, and tell him the path of gmond.conf.
The plugin is depend on libganglia.