write_tsdb: use different connection interval for each node
In ./src/write_tsdb.c, the wt_callback_init() is executed every time
collectd attempt to establish a connection to each database node.
So last_connect_timestamp and connect_interval must be declared and
manipulated separately for each node. Otherwise, when set to send data
to multiple node and failing to connect to one node, write_tsdb may skip
establishing a connection to the next node due to the interval, and this
is what actual code is.

Change-Id: If59efb05d60efa30a38155a1e33822f8eff7471b
Signed-off-by: envzhu
Signed-off-by: Li Xi <pkuelelixi@163.com>