filedata: ignore math failure
When trying to get client stats, got following error with info
log enabled:

failed to submit file data
unable to parse file /sys/kernel/debug/lustre/llite/lustre0-ffff8df476c94800/stats for type client_stats_link

This error message is not so helpful. In this patch, adds more
error messages for easier debug and ignore the math error.

The problem is caused by missing "^" in the pattern of
CLIENT_STATS_ITEM_FOUR/CLIENT_STATS_ITEM_FOUR_BYTES. The
pattern of link with matches the line of unlink and symlink.
That will cause duplicated value for math operations. And
another problem is the wrong/missing data of unlink and symlink.

Signed-off-by: Li Xi <pkuelelixi@163.com>